+++ assets/functions/templates/src/all-users.js
@@ -1,0 +1,14 @@
+async function allUsers() {
+  const { results } = await gql(`{
+    allUser {
+      results {
+        id
+        email
+      }
+    }
+  }`);
+
+  return results;
+}
+
+export { allUsers };