+++ assets/functions/templates/src/say-hello.js
@@ -1,0 +1,13 @@
+import join from 'lodash/join';
+
+async function sayHello() {
+  const name = await context('name');
+
+  if (name === 'oops') {
+    throw 'Ooops. Something went wrong.';
+  } else {
+    return join(['Hello', name], ',');
+  }
+}
+
+export { sayHello };